A0-1333 Reorganized code of performance page
